Adjusted a couple of variable names
Some changes to variable names to ensure compatibility between client/server and to aid readability. 

'response' to 'packet'. Response is used on the server side and has a specific meaning, the servers response back to the client. 'packet' seems appropriate, but you may have other preferences.

extended 'lat' and 'long' to their full names to be consistent with use elsewhere and on the server.

name clash between position as a parameter and position as a variable. Changed the parameter to 'location'. The code wasn't broken, it just makes reading it easier.

The timestamp, I have left this as is, but I'm not sure that we want to use the phones clock instead of the server. There may be a case for when the user is out of coverage, otherwise the server should perhaps be the default. Thinking further, the accuracy required by the park is likely to be 'give or take a day' as opposed to within a minute or two. In this case we could ignore the users timestamp totally. Its unlikely they will be out of coverage for more than a day.

Cheers!
Matt